rename cache util to cache io
