+++ protostar/self/cache_io.py
